removed routes
