+++ packages/backend/services/crm/index.ts
@@ -1,0 +1,8 @@
+export * from './contact';
+export * from './company';
+export * from './deal';
+export * from './event';
+export * from './lead';
+export * from './note';
+export * from './task';
+export * from './user';
